Create ProHKo
This workflow is designed to create protein HMMer profiles from KEGG Orthology database in an unsupervised manner. For the convenience of the user, the main function ProHKo will accept a KO number and the outputs will be the fasta sequences of the proteins in the same orthology group, their MSA and HMMer profiles.